yahtzee-back/users/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
black==23.12.1
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
cssbeautifier==1.14.11
djlint==1.34.1
dnspython==2.4.2
ecdsa==0.18.0
EditorConfig==0.12.3
exceptiongroup==1.2.0
fastapi==0.108.0
flake8==6.1.0
h11==0.14.0
html-tag-names==0.1.2
html-void-elements==0.1.0
httptools==0.6.1
idna==3.6
jsbeautifier==1.14.11
json5==0.9.14
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.1.0
pyasn1==0.5.1
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pyflakes==3.1.0
pymongo==4.6.1
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0.1
regex==2023.12.25
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.32.0.post1
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.9.0
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
